Health bar fix for GunDude desktop version
The healthbar display wasn't working right with the HealthBar.Width().

So I replaced it with the pixel width of the health bar when at 100% size. The number 216 is explained in the comment above this event, so people will hopefully not be confused by it if they want to alter it themselves in the future.